Move import logic from Participants to separate component
