+++ src/org/components/EventForm/steps/registration/Participants.tsx
@@ -1,14 +1,11 @@
 import { skipToken } from '@reduxjs/toolkit/dist/query'
 import { api } from 'app/services/bis'
 import type { User, UserPayload } from 'app/services/bisTypes'
-import spreadsheetTemplate from 'assets/templates/Vzor_import účastníků.xlsx'
 import classNames from 'classnames'
 import {
   Actions,
   Button,
   EmptyListPlaceholder,
-  ExternalButtonLink,
-  ImportExcelButton,
   Loading,
   SelectUnknownUser,
   StyledModal,
@@ -20,24 +17,14 @@ import {
   useShowMessage,
 } from 'features/systemMessage/useSystemMessage'
 import { merge } from 'lodash'
-import {
-  ConfirmedUser,
-  ImportParticipantsList,
-} from 'org/components/ImportParticipantsList/ImportParticipantsList'
+import { ImportParticipants } from 'org/components/ImportParticipants/ImportParticipants'
+import { ConfirmedUser } from 'org/components/ImportParticipants/ImportParticipantsList/ImportParticipantsList'
 import { FC, useState } from 'react'
 import { FaTrash as Bin, FaUserEdit as EditUser } from 'react-icons/fa'
 import colors from 'styles/colors.module.scss'
-import { Overwrite } from 'utility-types'
 import { ApplicationStates } from '../ParticipantsStep'
 import styles from '../ParticipantsStep.module.scss'
 import { ShowApplicationModal } from './ShowApplicationModal'
-
-// type UserImport = Omit<UserPayload, 'pronoun' | 'all_emails'>
-
-export type UserImport = Overwrite<
-  Omit<UserPayload, 'pronoun' | 'subscribed_to_newsletter' | 'all_emails'>,
-  { health_insurance_company: string; address: string; contact_address: string }
->
 
 export const Participants: FC<{
   eventId: number
@@ -246,21 +233,6 @@ export const Participants: FC<{
     }
   }
 
-  const [importParticipantsModalOpen, setImportParticipantsModalOpen] =
-    useState(false)
-  const [importParticipantsData, setImportParticipantsData] =
-    useState<UserImport[]>()
-
-  const handleCancelImportParticipants = () => {
-    setImportParticipantsModalOpen(false)
-    setImportParticipantsData(undefined)
-  }
-
-  const handleImportParticipants = async (data: UserImport[]) => {
-    setImportParticipantsData(data)
-    setImportParticipantsModalOpen(true)
-  }
-
   const handleSaveImportedParticipants = async (data: ConfirmedUser[]) => {
     const existingParticipants = data.filter(
       datum => 'id' in datum && datum.id,
@@ -287,8 +259,6 @@ export const Participants: FC<{
         await updateUser({ id, patchedUser: data })
       }),
     )
-
-    setImportParticipantsModalOpen(false)
   }
 
   const removeModalTitle = removeModalData
@@ -319,46 +289,9 @@ export const Participants: FC<{
           onCancel={handleCancelRemoveParticipant}
         />
       </StyledModal>
-      <StyledModal
-        title="Import účastnic a účastníků"
-        open={importParticipantsModalOpen}
-        onClose={handleCancelImportParticipants}
-      >
-        <ImportParticipantsList
-          data={importParticipantsData}
-          onConfirm={handleSaveImportedParticipants}
-          onCancel={handleCancelImportParticipants}
-        />
-      </StyledModal>
+
       <h2>Účastníci</h2>
-      <ImportExcelButton<UserImport>
-        keyMap={{
-          first_name: 0,
-          last_name: 1,
-          birthday: 2,
-          email: 3,
-          phone: 4,
-          address: 5,
-          contact_address: 6,
-          health_insurance_company: 7,
-          health_issues: 8,
-          close_person: {
-            first_name: 9,
-            last_name: 10,
-            email: 11,
-            phone: 12,
-          },
-          birth_name: 13,
-          nickname: 14,
-        }}
-        headerRows={2}
-        onUpload={handleImportParticipants}
-      >
-        Importovat z excelu
-      </ImportExcelButton>
-      <ExternalButtonLink tertiary href={spreadsheetTemplate}>
-        (vzor)
-      </ExternalButtonLink>
+      <ImportParticipants onConfirm={handleSaveImportedParticipants} />
       {!isReadParticipantsLoading ? (
         <div>
           <div>Přidat účastníka:</div>
