+++ src/org/components/ImportParticipants/ImportParticipants.tsx
@@ -1,0 +1,89 @@
+import { UserPayload } from 'app/services/bisTypes'
+import spreadsheetTemplate from 'assets/templates/Vzor_import účastníků.xlsx'
+import { ExternalButtonLink, ImportExcelButton, StyledModal } from 'components'
+import { useState } from 'react'
+import { Overwrite } from 'utility-types'
+import {
+  ConfirmedUser,
+  ImportParticipantsList,
+} from './ImportParticipantsList/ImportParticipantsList'
+
+export type UserImport = Overwrite<
+  Omit<UserPayload, 'pronoun' | 'subscribed_to_newsletter' | 'all_emails'>,
+  {
+    health_insurance_company: string
+    address: string
+    contact_address: string
+  }
+>
+
+export const ImportParticipants = ({
+  onConfirm,
+}: {
+  onConfirm: (data: ConfirmedUser[]) => Promise<void>
+}) => {
+  const [importParticipantsModalOpen, setImportParticipantsModalOpen] =
+    useState(false)
+
+  const [importParticipantsData, setImportParticipantsData] =
+    useState<UserImport[]>()
+
+  const handleConfirm = async (data: ConfirmedUser[]) => {
+    await onConfirm(data)
+    setImportParticipantsModalOpen(false)
+  }
+
+  const handleCancel = () => {
+    setImportParticipantsModalOpen(false)
+    setImportParticipantsData(undefined)
+  }
+
+  const handleImportParticipants = async (data: UserImport[]) => {
+    setImportParticipantsData(data)
+    setImportParticipantsModalOpen(true)
+  }
+
+  return (
+    <>
+      <StyledModal
+        title="Import účastnic a účastníků"
+        open={importParticipantsModalOpen}
+        onClose={handleCancel}
+      >
+        <ImportParticipantsList
+          data={importParticipantsData}
+          onConfirm={handleConfirm}
+          onCancel={handleCancel}
+        />
+      </StyledModal>
+      <ImportExcelButton<UserImport>
+        keyMap={{
+          first_name: 0,
+          last_name: 1,
+          birthday: 2,
+          email: 3,
+          phone: 4,
+          address: 5,
+          contact_address: 6,
+          health_insurance_company: 7,
+          health_issues: 8,
+          close_person: {
+            first_name: 9,
+            last_name: 10,
+            email: 11,
+            phone: 12,
+          },
+          birth_name: 13,
+          nickname: 14,
+        }}
+        headerRows={2}
+        onUpload={handleImportParticipants}
+      >
+        Importovat z excelu
+      </ImportExcelButton>
+      <ExternalButtonLink tertiary href={spreadsheetTemplate}>
+        (vzor)
+      </ExternalButtonLink>
+    </>
+  )
+}
